Don't add IProjectQueryService to DI in the language server
