+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Extensions/IServiceCollectionExtensions.cs
@@ -226,7 +226,6 @@ internal static class IServiceCollectionExtensions
         // Add project snapshot manager
         services.AddSingleton<IProjectEngineFactoryProvider, LspProjectEngineFactoryProvider>();
         services.AddSingleton<IProjectSnapshotManager, LspProjectSnapshotManager>();
-        services.AddSingleton<IProjectQueryService, LspProjectQueryService>();
     }
 
     public static void AddHandlerWithCapabilities<T>(this IServiceCollection services)
